Слил PostStatusController и PostStatusService соответсвенно в PostController и PostService
