+++ infosync-app/src/main/java/ru/urfu/infosync/service/PostService.java
@@ -3,6 +3,8 @@ package ru.urfu.infosync.service;
 import org.springframework.stereotype.Service;
 import ru.urfu.infosync.dao.PostDao;
 import ru.urfu.infosync.model.GeneralPost;
+import ru.urfu.infosync.model.PostStatus;
+import ru.urfu.infosync.dao.PostStatusDao;
 
 import java.util.List;
 
@@ -15,13 +17,19 @@ public class PostService {
 
     private final PostDao postDao;
     private final UserService userService;
+    private final PostStatusDao postStatusDao;
 
-    public PostService(final PostDao postDao, final UserService userService) {
+    public PostService(final PostDao postDao, final UserService userService, final PostStatusDao postStatusDao) {
         this.postDao = postDao;
         this.userService = userService;
+        this.postStatusDao = postStatusDao;
     }
 
     public List<GeneralPost> getNews() {
         return postDao.getRecommendedPosts(userService.getCurrentUser().getGroupId());
+    }
+
+    public void markAsRead (Integer postId) {
+        postStatusDao.setPostStatusTrue(new PostStatus(userService.getIdCurrentUser(), postId));
     }
 }
